requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
pdfkit==1.0.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
stripe==10.3.0
typing_extensions==4.12.2
urllib3==2.2.2
